test(history): consolidate tool-execution-logger tests under test/subscribers
Two same-named suites tested src/subscribers/tool-execution-logger.ts —
one misplaced under test/services/. Consolidate into the canonical
test/subscribers/ location (mirroring src/) to avoid drift.

Ported the cases unique to the services suite before deleting it, so no
coverage is lost: google_search (query) and fetch_url (url) key-param
mapping, merging into a header-only tools callout, and multi-line merges.
